+++ app/models/data_availability.rb
@@ -1,0 +1,2 @@
+class DataAvailability < Struct.new(:id, :description, :available?)
+end