+++ app/presenters/teacher_survey_presenter.rb
@@ -1,6 +1,6 @@
 class TeacherSurveyPresenter < DataItemPresenter
-  def initialize(measure_id:, survey_items:, has_sufficient_data:)
-    super(measure_id:, has_sufficient_data:)
+  def initialize(measure_id:, survey_items:, has_sufficient_data:, school:, academic_year:)
+    super(measure_id:, has_sufficient_data:, school:, academic_year:)
     @survey_items = survey_items
   end
 
@@ -20,5 +20,11 @@ class TeacherSurveyPresenter < DataItemPresenter
 
   def reason_for_insufficiency
     'low response rate'
+  end
+
+  def descriptions_and_availability
+    @survey_items.map do |survey_item|
+      DataAvailability.new(survey_item.survey_item_id, survey_item.prompt, true)
+    end
   end
 end