+++ spec/services/admin_data_loader_spec.rb
@@ -5,6 +5,7 @@ describe AdminDataLoader do
   let(:ay_2018_19) { AcademicYear.find_by_range '2018-19' }
   let(:attleboro) { School.find_by_dese_id 160_505 }
   let(:winchester) { School.find_by_dese_id 3_440_505 }
+  let(:beachmont) { School.find_by_dese_id 2_480_013 }
   let(:chronic_absense_rate) { AdminDataItem.find_by_admin_data_item_id 'a-vale-i1' }
   let(:student_to_instructor_ratio) { AdminDataItem.find_by_admin_data_item_id 'a-sust-i3' }
 
@@ -28,7 +29,7 @@ describe AdminDataLoader do
 
       # it 'assigns the school to the admin data value' do
       expect(AdminDataValue.first.school).to eq attleboro
-      expect(AdminDataValue.last.school).to eq winchester
+      expect(AdminDataValue.last.school).to eq beachmont
       # end
 
       # it 'links the admin data value to the correct admin data item' do
@@ -37,12 +38,14 @@ describe AdminDataLoader do
       # end
 
       # it 'loads all the admin data values in the target csv file' do
-      expect(AdminDataValue.count).to eq 11
+      expect(AdminDataValue.count).to eq 10
       # end
 
       # it 'captures the likert score ' do
-      expect(AdminDataValue.first.likert_score).to eq 3.03
-      expect(AdminDataValue.last.likert_score).to eq 5
+      expect(AdminDataValue.find_by(school: attleboro, academic_year: ay_2018_19,
+                                    admin_data_item: chronic_absense_rate).likert_score).to eq 3.03
+      expect(AdminDataValue.find_by(school: beachmont, academic_year: ay_2018_19,
+                                    admin_data_item: student_to_instructor_ratio).likert_score).to eq 3.5
       # end
 
       # it 'rounds up any likert_scores between 0 and 1 (non-inclusive) to 1' do
@@ -52,7 +55,7 @@ describe AdminDataLoader do
       # it 'rejects importing rows with a value of 0' do
       expect(AdminDataValue.where(school: attleboro, academic_year: ay_2018_19,
                                   admin_data_item: AdminDataItem.find_by_admin_data_item_id('a-reso-i1'))).not_to exist
-      expect(AdminDataValue.where(school: attleboro, academic_year: ay_2018_19,
+      expect(AdminDataValue.where(school: winchester, academic_year: ay_2018_19,
                                   admin_data_item: AdminDataItem.find_by_admin_data_item_id('a-sust-i3'))).not_to exist
       # end
     end